gas: Move gas module from runtime to graph crate
This will probably be necessary for multiblockchain.
