+++ runtime/wasm/src/lib.rs
@@ -16,7 +16,7 @@ pub mod host_exports;
 
 pub mod error;
 
-mod gas;
+mod gas_rules;
 
 pub use host::RuntimeHostBuilder;
 pub use host_exports::HostExports;
