+++ runtime/wasm/src/mapping.rs
@@ -1,10 +1,11 @@
-use crate::gas::{Gas, GasRules};
+use crate::gas_rules::GasRules;
 use crate::module::{ExperimentalFeatures, WasmInstance};
 use futures::sync::mpsc;
 use futures03::channel::oneshot::Sender;
 use graph::blockchain::{Blockchain, HostFn, TriggerWithHandler};
 use graph::components::subgraph::{MappingError, SharedProofOfIndexing};
 use graph::prelude::*;
+use graph::runtime::gas::Gas;
 use std::collections::BTreeMap;
 use std::sync::Arc;
 use std::thread;
